Allow assigning meeting license on synced users
Add new configuration property that specifies a Guid that represents a
valid TeamViewer Meeting license activation code. If set, the license
will automatically be activated on newly created company members.

Note: For this to work the license must already be activated on the
company profile and it must still have free organizer license slots,
else the set license will be silently ignored.

* Add optional 'MeetingLicenseKey' property to configuration.
* Forward set meeting license key to create member Web API call